+++ app/src/services/file/file.functions.ts
@@ -3,8 +3,8 @@
  */
 import { createServerFn } from "@tanstack/react-start";
 import { z } from "zod";
-import { PERMISSIONS } from "#/constants/permissions/permissions";
 import { adminPermGuard } from "#/middleware/admin-auth";
+import { PERMISSIONS } from "#/permissions/permissions";
 import { logCrud } from "#/services/operation-log/operation-log.server";
 import { getFileInfo, getFileList, uploadFile } from "./file.server";